feat(post): enable clickable username and optimize author link hover interaction

+++ dao/mysql/post.go
@@ -15,7 +15,7 @@ func CreatePost(p *models.Post) (err error) {
 	if res.Error != nil {
 		zap.L().Error("create post failed",
 			zap.String("operation", "create_post"),
-			zap.Int64("author_id", p.AuthorID),
+			zap.Int64("author_id", int64(p.AuthorID)),
 			zap.Int64("community_id", p.CommunityID),
 			zap.Error(res.Error))
 		return res.Error
